Consider only active workers as quorum members
If a worker connects and then disconnects while there are still tests
left to run we don't want to count it as a member of a quorum. What we
really care about is "do we have enough workers to run the remaining
tests?", which we're fuzzily answering by giving ourselves M seconds to
get at least N workers connected.