Test the central master's exit code, not the remote master's
In these tests, only a single test fails. That failing test could be run
on either the central master or the remote master. But regardless of
which master it runs on, the central master should fail, so that's the
process whose exit code we should be testing.

Note that the remote master now starts up first. It shouldn't matter
which master starts first (and ideally we'd have tests for both orders)
and doing it this way keeps the test easier to read.