Fix test queue resultes for Minitest 5
As noted in issue #22 Minitest 5 changed the results so that outputting
the test queue results in `record` causes the suite to be output as
such:

```
SuiteName: .
SuiteName: .
SuiteName: .
```

The output should instead look like:

```
SuiteName: ...
```

This commits collects the results in `record` and then groups those
results by class so it can iterate through each grouped class and
properly report pass/skip/fail via test queue output.

We need to use the SummaryReporter here instead of the ProgressReporter
so we have access to `results` in `record`.

Closes #22