Move bindir from bin directory to exe directory
As shown in the template `bundle gem` below, it seems that it is
main stream to arrange bindir as exe directory instead of bin directory.
https://github.com/bundler/bundler/blob/v1.16.2/lib/bundler/templates/newgem/newgem.gemspec.tt#L36-L37

This PR moves executable commands from bin directory to exe directory.

Placing development commands in the bin directory makes it easier to
understand as it becomes a directory structure according to main stream.