Run with `spec` directory by default for `rspec-queue` command when using RSpec 3
## Before

Requrie a target target directory:

```console
$ bundle exec test-queue spec
```

If the `spec` path is omitted, user might be surprised to see 0 test runs.

## After

`spec` directory is target by default when target directory is ommited:

```console
$ bundle exec test-queue
```

So, user can use it in the same abbreviated way as RSpec.

```console
$ bundle exec rspec # spec directory by default target.
```

Note, however, that this PR implementation (i.e. `@files_or_directories_to_run`) is naive as it relies on:
https://github.com/rspec/rspec-core/blob/v3.12.1/lib/rspec/core/configuration.rb#L1061-L1077

Note, it does not support legacy RSpec 2.
So rspec.bats is copied to rspec2.bats and rspec3.bats, and adding new test case only to rspec3.bats.