Drop support for Cucumber v1.3
This commit removes support for Cucumber v1.3.

The last release of Cucumber v1.3 was nearly 10 years ago:
https://github.com/cucumber/cucumber-ruby/releases/tag/v1.3.20
It's hard to imagine anyone using such an old version of Cucumber with test-queue in 2025.

I’m also working on adding Ruby 3.4 support to CI for test-queue,
but running test-queue with Cucumber v1.3 on Ruby 3.4 is problematic:
https://github.com/S-H-GAMELINKS/test-queue/pull/1

Given these factors, I think it’s reasonable to drop support for Cucumber v1.3.